✨feat: add emoji to commit message
Adds an optional `--emoji` flag to the `commit` command that will add an emoji to the commit message.

The emoji will be chosen based on the type of commit message. For example, a feat commit will have a ✨ emoji, a fix commit will have a 🚑 emoji, and a docs commit will have a 📝 emoji.